Refactor setting dialog view and viewmodel - extract tab views and viewmodel from the dialog into separate files
