+++ Up2dateService/Up2dateConsole/Dialogs/Settings/ConsoleSecurityTab.xaml
@@ -1,0 +1,22 @@
+﻿<UserControl x:Uid="UserControl_1" x:Class="Up2dateConsole.Dialogs.Settings.ConsoleSecurityTab"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Up2dateConsole.Dialogs.Settings"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="500"
+             d:DataContext="{d:DesignInstance local:ConsoleSecurityTabViewModel}">
+    
+    <StackPanel x:Uid="StackPanel_7">
+        <CheckBox x:Uid="CheckBox_3" Margin="0,15,0,5" Content="Leave administrative mode on console close" 
+                              IsChecked="{Binding LeaveAdminModeOnClose, UpdateSourceTrigger=PropertyChanged}"/>
+        <CheckBox x:Uid="CheckBox_4" Margin="0,10,0,5" Content="Leave administrative mode when console is not used for" 
+                              IsChecked="{Binding LeaveAdminModeOnInactivity, UpdateSourceTrigger=PropertyChanged}"/>
+        <StackPanel x:Uid="StackPanel_8" Margin="20,0,0,5" Orientation="Horizontal">
+            <TextBox x:Uid="TextBox_3" Width="40" HorizontalAlignment="Left" Text="{Binding LeaveAdminModeOnInactivityTimeout, UpdateSourceTrigger=PropertyChanged}"
+                                 IsEnabled="{Binding LeaveAdminModeOnInactivity}"/>
+            <TextBlock x:Uid="TextBlock_5" Margin="8,0" Text="seconds"/>
+        </StackPanel>
+    </StackPanel>
+</UserControl>
