+++ Up2dateService/Up2dateConsole/Dialogs/Settings/ServerConnectionTab.xaml
@@ -1,0 +1,18 @@
+﻿<UserControl x:Uid="UserControl_1" x:Class="Up2dateConsole.Dialogs.Settings.ServerConnectionTab"
+             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
+             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
+             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
+             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
+             xmlns:local="clr-namespace:Up2dateConsole.Dialogs.Settings"
+             mc:Ignorable="d" 
+             d:DesignHeight="300" d:DesignWidth="500"
+             d:DataContext="{d:DesignInstance local:ServerConnectionTabViewModel}">
+    
+    <StackPanel x:Uid="StackPanel_5">
+        <TextBlock x:Uid="TextBlock_1" Margin="0,15,0,5" Text="URL for getting authorization certificate by one-time key:"/>
+        <TextBox x:Uid="TextBox_1" Margin="0,0,0,5" Text="{Binding TokenUrl, UpdateSourceTrigger=PropertyChanged}"/>
+        <TextBlock x:Uid="TextBlock_2" Margin="0,5,0,5" Text="Provisioning URL:"/>
+        <TextBox x:Uid="TextBox_2" Margin="0,0,0,5" Text="{Binding DpsUrl, UpdateSourceTrigger=PropertyChanged}"/>
+    </StackPanel>
+    
+</UserControl>
