+++ Up2dateService/Up2dateConsole/Dialogs/Settings/SettingsDialog.xaml.cs
@@ -1,4 +1,4 @@
-﻿namespace Up2dateConsole.Dialogs
+﻿namespace Up2dateConsole.Dialogs.Settings
 {
     public partial class SettingsDialog
     {
