+++ Up2dateService/Up2dateConsole/Texts.cs
@@ -56,6 +56,9 @@
         PackageStatusAccepting,
         PackageStatusRejecting,
         CannotStopService,
-        CannotStartService
+        CannotStartService,
+        ServerConnection,
+        InstallationPolicy,
+        ConsoleSecurity
     }
 }
