+++ Up2dateService/Up2dateConsole/Up2dateConsole.csproj
@@ -75,14 +75,26 @@
       <DependentUpon>Reference.svcmap</DependentUpon>
     </Compile>
     <Compile Include="Controls\SqueezeContentControl.cs" />
+    <Compile Include="Dialogs\Settings\ConsoleSecurityTab.xaml.cs">
+      <DependentUpon>ConsoleSecurityTab.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Dialogs\Settings\ConsoleSecurityTabViewModel.cs" />
+    <Compile Include="Dialogs\Settings\InstallationPolicyTab.xaml.cs">
+      <DependentUpon>InstallationPolicyTab.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Dialogs\Settings\InstallationPolicyTabViewModel.cs" />
     <Compile Include="Dialogs\RequestCertificateDialogViewModel.cs" />
     <Compile Include="Dialogs\RequestCertificateDialog.xaml.cs">
       <DependentUpon>RequestCertificateDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Dialogs\SettingsDialog.xaml.cs">
+    <Compile Include="Dialogs\Settings\ServerConnectionTab.xaml.cs">
+      <DependentUpon>ServerConnectionTab.xaml</DependentUpon>
+    </Compile>
+    <Compile Include="Dialogs\Settings\ServerConnectionTabViewModel.cs" />
+    <Compile Include="Dialogs\Settings\SettingsDialog.xaml.cs">
       <DependentUpon>SettingsDialog.xaml</DependentUpon>
     </Compile>
-    <Compile Include="Dialogs\SettingsDialogViewModel.cs" />
+    <Compile Include="Dialogs\Settings\SettingsDialogViewModel.cs" />
     <Compile Include="Helpers\CommandLineHelper.cs" />
     <Compile Include="Helpers\InactivityMonitor\IInactivityMonitor.cs" />
     <Compile Include="Helpers\InactivityMonitor\HookMonitor.cs" />
@@ -118,7 +130,19 @@
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
-    <Page Include="Dialogs\SettingsDialog.xaml">
+    <Page Include="Dialogs\Settings\ConsoleSecurityTab.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Dialogs\Settings\InstallationPolicyTab.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Dialogs\Settings\ServerConnectionTab.xaml">
+      <SubType>Designer</SubType>
+      <Generator>MSBuild:Compile</Generator>
+    </Page>
+    <Page Include="Dialogs\Settings\SettingsDialog.xaml">
       <SubType>Designer</SubType>
       <Generator>MSBuild:Compile</Generator>
     </Page>
@@ -258,6 +282,7 @@
   <ItemGroup>
     <WCFMetadataStorage Include="Connected Services\ServiceReference\" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <PropertyGroup>
     <PostBuildEvent>"$(ProjectDir)Localization\CheckAndGenerate.cmd" ru-RU $(TargetDir) $(TargetName) $(ProjectDir) $(ProjectName)
